add skeleton of rainbow
This is a basic setup that has a grpc based server and
client, with the intention to have simple endpoints
to register clusters, and then to submit jobs (from
either a cluster or a standalone client) and then poll
for jobs (from a registered cluster). Next I am going to
add the little sqlite database to enforce the register
command, and then likely work on the submit job command
and adding the secret / handshake bit.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>