feat: support for endpoint to request jobs
With this addition, a cluster can provide the secret given to it upon registration to receive a max number of jobs. There is no obligation to run them, but
rather we next will implement a response that says "I have accepted these
ids" and then the rainbow scheduler will delete them from its database,
as it is just acting as a dispatcher.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>