add docker-compose example
This demo shows bringing up the rainbow scheduler and three clusters,
each with flux, and via containers, and then having each register
to the rainbow "scheduler" (dispatcher) and randomly submit and
receive jobs, from any cluster (including themselves). It randomly
selects N for submit and selection and runs for 10 iterations. It is
the most basic demo for a "multi-cluster" approach using flux. We
next need to add the actual scheduler :)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>