wip: preparing to integrate scheduler
Problem: we need to be able to submit something that is more akin
to a jobspec or container that has compatibility artifacts, and have
the job matched to one of several clusters.
Solution: first allow a file with cluster credentials (mostly to make
life easier) and then the register endpoint should send over cluster
metadata. In order to do this we should extract metadata from compspec
(for one node) into some directory, and then have a summary function
to generate a resource specification to describe a cluster. I am
switching over to compspec go to work on that, and then will return
here. The node should have this directory or the single cluster
file prepared before registration.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>