feat: add support for loading graph
Problem: we need the register step to send over a JGF graph
Solution: serialize as a string and load into the server.

I need to decide what to do with this next. Ideally I could
use fluxion, but it does not support updating the main graph
once it is made, so maybe I will prototype something else
in the  meantime.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>