feat: add in memory graph database
Problem: we need a simple prototype for an in memory graph database
Solution: make one! This is almost done - I have the graph loading into
the MemoryGraph and just need to make the vertices and edges. I think to
start for multiple graphs I am just going to put them at a common root.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>