feat: add support for basic subsystem
Problem: we will eventually want to have multiple subsystems
Solution: for the time being, create a cluster graph with
a named subsystem lookup. Each item in the lookup is a named
subsystem, with a dominant one defined on the initial cluster
graph creation (currently set to be "nodes"). We can then
register clusters to the dominant subsystem (and eventually
register or add metadata for other subsystems). For an
early prototype, each subsystem graph will have a common
root, off of which named clusters will be defined. Metrics
will be added for each graph under the subsystem for a quick
assessment if some total resource is sufficient to warrant
traversing the graph (not done yet)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>