feat: support for basic satisfy logic
Problem: we want to be able to submit a jobspec and do a satisfy request against the graph.
Solution: this WIP adds the parsing of a jobspec (from simple parameters into the Jobspec,
which is a library proper, and in the future we can have converters from other kinds of
job files or even Kubernetes abstractions). The Jobspec goes from the calling client directly
to the graph database, which is a GRPC endpoint, and that logic is exposed from the graph
database interface. Since this is an in-memory database provided directly by rainbow,
this means it is technically still hitting the rainbow server, but in the future this can
be a separate graph database GRPC endpoint. Note that auth will need to be worked out for
when the graph is separate - likely when a cluster registers it will have permission/auth
as part of the schema. As is implemented here, we then hit the actual in memory graph,
and run a Satisfy -> DFSForMatch function, which right now is just doing a top level
check for resources of each known cluster against the Jobspec. I am doing this
because I do not yet know what the DFS should look like (I need to look at Fluxion code)
but the response returns a simple list of clusters that can satisfy the resources from
a global (totals) sense. We also need to think about what information about the match
(beyond the cluster name) should go back to the calling client to be sent to rainbow.
Note that the client is handling these two requests, and the user is just making
one request with their job.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>