feat: add support for assignment table
When the graph database returns clusters that satisfy a jobspec, they need to be redirected
to the rainbow cluster to be assigned. This is a two step process, where first we add
the jobid to the jobs table (and it is not assigned) and then we will add it to an assignment
table with each cluster id that can receive it. At this point we are going to ask the clusters
if they can satisfy (and when) to assign it, and I do think we need a push model for the assignment.
We will need to allow for failure to connect (and retry) and some kind of heartbeat to do that,
but I first need to think about how a cluster can have work pushed to it - likely we need a
special client running there that I have not written yet

Signed-off-by: vsoch <vsoch@users.noreply.github.com>