feat: add support for selection algorithm plugins
Problem: the current approach to receive a list of clusters that satisfy
and then need to ask each one for more informaiton is problematic - it
could lead to an inability to reach consensus. Instead, we can provide
the list of satisfied clusters to rainbow, and rainbow can use some (TBA)
cluster metrics (that are updated) that reflect state to do a single
assignment. This means that only one pull is needed, and the jobs
received are decidedly assigned to that cluster. It essentially
removes an entire step! To support selection algorithms as another
plugin type I have also reorganized the backends into a plugins folder
so the plugin functionality is clear.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>