feat: receive jobs
The new receive jobs endpoint is a request made by a cluster
to receive jobs. The jobs have already been assigned to it,
so despite the pull model (which seems necessary since we likely
will not have clusters exposing their own API endpoints) the
cluster needs to have some kind of binding contract with rainbow
that it indeed will do the pull to receive the jobs. What I have
not done yet is the python cases implementation, and I can do that
next, followed by prototyping alongside flux and then thinking
about adding a subsystem.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>