start of thinking for subsystem design
Problem: we need to be able to register a subsystem
Solution: define the subsystem as its own graph, where
the root is a handle to the children, the graph
itself is namespaced to the subsystem, and it is required
that each child (resource type) can be linked to a node
in the dominant subsystem graph. It is also not allowed
to add a node type that is not currently present in the
dominant subsystem. This is just the design for the json
and I will work on the register logic (and validation)
next.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>