feat: finishing subsystem register to add match
Problem: subsystem resources need to be accounted for during a match
Solution: Since each subsystem vertex (e.g., an io kind of storage) has a bidirectional
link to a node in the dominant subsystem, and the jobspec need for a subsystem resource
is defined on the level of a slot, we can very easily add logic to the findSlots recursive
function that accounts for if the slot also satisfies all the features that are required
by the resources section. To implement this, I created a simple design under the slot
for "resources" that are namespaced by the subsystem, and then for a "first shot" algorithm
- "match" is basically going to look for matching key value pairs. Given that all key value
pairs are satisfied for a slot, we can consider the slot a match and return it. That level
of logic can be extended to do other kinds of assessment of the subsystem, either taking
into account a count or even traversing the subsystem graph from that node, but we do
not need that level of complexity for these first level scheduling experiments, and I want
to work on other things, so I am leaving it at that.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>