feat: support for selection and match algorithms
Problem: we currently have an interface to selection but not match algorithms
Solution: improve the algorithm interface by removing the match logic from
the graph, and creating an interface that the graph uses. In practice I wanted
to use this when I realized I needed a new match algorithm to match in a range of
strings.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>