feat: state endpoint
Problem: we want to be able to select a final cluster based on state
Solution: while a cluster will likely send back state when accepting jobs,
for the time being we are providing a one off endpoint for a cluster to
send state data, and it requires the cluster to authenticate with the name
and secret. For now the state date is key value pairs of compatibility
metadata, and the values can be anything (interface) as we are expecting
the selection algorithms that use them to know the type for a specific key.
I will be added selection algorithms after this first stage that I intend
to use in scheduler simulation - namely a post filter that serves as
additional constraints that can be chosen (max jobs a cluster can accept
and a cost value that can be minimized.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>