bug: matches not accounted for with go client
Problem: we needed to take the intersection of the
matches and clusters known to the user config as
the final set to return. Currently we are just
providing all the clusters in the user config,
regardless of match.
Solution: create a function to retrieve the token
for a named cluster, which only returns the token
if it is known. This is akin to taking an intersection
between the matches and clusters known.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>