feat: add metrics and timing functions
Problem: we need to record timing of things
Solution: this adds a metrics table to the database, which
is accessible directly by the same rainbow server that holds the
database, and otherwise (for clients that are not directly connected)
it can be sent metrics to save. Saving means a name, value, and arbitrary
set of key value pairs (both strings). I am also including a pythons
script that shows how to read the rainbow database and output the
tables.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>